requirements.txt
# MoxNAS Backend Requirements
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
psutil==5.9.6
paramiko==3.3.1
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
proxmoxer==2.0.1

# Development dependencies
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0
